% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-append.R
\name{ec_append}
\alias{ec_append}
\title{Append a data frame to a "room", optionally starting at a given row}
\usage{
ec_append(x, room, at = "end", browse = FALSE, ...,
  ec_host = ethercalc_host())
}
\arguments{
\item{x}{a data frame}

\item{room}{name of an existing EtherCalc "room".}

\item{at}{if not "\code{end}" then the row where the data will be inserted and pasted at.}

\item{browse}{if \code{TRUE} (\code{FALSE} is the default) then open up a browser tab/window to
the room}

\item{...}{passed on to \code{\link[=write.csv]{write.csv()}}}

\item{ec_host}{See \code{\link[=ethercalc_host]{ethercalc_host()}}}
}
\value{
the commands used by EtherCalc to process the append operation (invisibly)
}
\description{
Append a data frame to a "room", optionally starting at a given row
}
\examples{
ec_edit(iris, "test")
ec_append(iris, "test")
}
