% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-cmd.R
\name{ec_cmd}
\alias{ec_cmd}
\title{Issue EtherCalc "commands" to a room}
\usage{
ec_cmd(commands = "", room, ec_host = ethercalc_host())
}
\arguments{
\item{commands}{character vector of valid EtherCalc commands}

\item{room}{name of an EtherCalc "room"}

\item{ec_host}{See \code{\link[=ethercalc_host]{ethercalc_host()}}}
}
\description{
Issue EtherCalc "commands" to a room
}
\examples{
ec_cmd(c("set A1:E1 font normal bold \\* \\*", "deleterow A12"), "test")
}
