% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-delete.R
\name{ec_delete}
\alias{ec_delete}
\title{Delete a "room" from an EtherCalc instance}
\usage{
ec_delete(room, ec_host = ethercalc_host())
}
\arguments{
\item{room}{name of an EtherCalc "room"}

\item{ec_host}{See \code{\link[=ethercalc_host]{ethercalc_host()}}}
}
\description{
Delete a "room" from an EtherCalc instance
}
\examples{
ec_delete("test")
}
