% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-edit.R
\name{ec_edit}
\alias{ec_edit}
\title{Edit a data frame in a new or existing EtherCalc "room"}
\usage{
ec_edit(x, room = NULL, browse = TRUE, ...,
  ec_host = ethercalc_host())
}
\arguments{
\item{x}{a data frame}

\item{room}{name of an existing EtherCalc "room" or \code{NULL} to create a new one}

\item{browse}{if \code{TRUE} (default) then open up a browser tab/window to the new room}

\item{...}{passed on to \code{\link[=write.csv]{write.csv()}}}

\item{ec_host}{See \code{\link[=ethercalc_host]{ethercalc_host()}}}
}
\description{
Edit a data frame in a new or existing EtherCalc "room"
}
\examples{
ec_edit(iris, "test")
}
