% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-exists.R
\name{ec_exists}
\alias{ec_exists}
\title{Does an EtherCalc "room" name exist?}
\usage{
ec_exists(room, ec_host = ethercalc_host())
}
\arguments{
\item{room}{name of an EtherCalc "room"}

\item{ec_host}{See \code{\link[=ethercalc_host]{ethercalc_host()}}}
}
\description{
Does an EtherCalc "room" name exist?
}
\examples{
\dontrun{
ec_exists("test")
}
}
