% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-export.R
\name{ec_export}
\alias{ec_export}
\title{Export an EtherCalc "room" to a local data frame}
\usage{
ec_export(room, ec_host = ethercalc_host())
}
\arguments{
\item{room}{name of an EtherCalc "room"}

\item{ec_host}{See \code{\link[=ethercalc_host]{ethercalc_host()}}}
}
\description{
Export an EtherCalc "room" to a local data frame
}
