% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-list.R
\name{ec_list}
\alias{ec_list}
\title{List all of the available EtherCalc documents}
\usage{
ec_list(all = FALSE, ec_host = ethercalc_host())
}
\arguments{
\item{all}{if \code{TRUE} then return a list of all "rooms" including the
internal-use only \code{_formdata} rooms.}

\item{ec_host}{See \code{\link[=ethercalc_host]{ethercalc_host()}}}
}
\description{
Connects to the \code{ec_host} EtherCalc instance and retrieves
a list of all the available documents. Note that documents
ending in \code{_formdata} are for internal API use only and should
not be used as "room" names.
}
\examples{
\dontrun{
ec_list()
}
}
