% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-new.R
\name{ec_new}
\alias{ec_new}
\title{Create a new, empty EtherCalc "room"}
\usage{
ec_new(room, browse = TRUE, ec_host = ethercalc_host())
}
\arguments{
\item{room}{name of an EtherCalc "room"}

\item{browse}{if \code{TRUE} (default) then open up a browser tab/window to the new room}

\item{ec_host}{See \code{\link[=ethercalc_host]{ethercalc_host()}}}
}
\value{
the full URL to the newly created room (invisibly)
}
\description{
Create a new, empty EtherCalc "room"
}
\examples{
\dontrun{
ec_new("test")
}
}
