% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-export.R
\name{ec_read}
\alias{ec_read}
\title{Read an EtherCalc "room" to a local data frame}
\usage{
ec_read(room, ..., ec_host = ethercalc_host())
}
\arguments{
\item{room}{name of an EtherCalc "room"}

\item{...}{passed on to \code{\link[readr:read_csv]{readr::read_csv()}}}

\item{ec_host}{See \code{\link[=ethercalc_host]{ethercalc_host()}}}
}
\description{
Read an EtherCalc "room" to a local data frame
}
\examples{
\dontrun{
ec_edit(mtcars, "mtcars")
ec_read("mtcars")
}
}
\seealso{
Other EthercCalc importers/exporters: \code{\link{ec_append}},
  \code{\link{ec_edit}}, \code{\link{ec_export}}
}
\concept{EthercCalc importers/exporters}
