% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-new.R
\name{ec_view}
\alias{ec_view}
\title{Open a browser window/tab to an EtherCalc "room"}
\usage{
ec_view(room, ec_host = ethercalc_host())
}
\arguments{
\item{room}{name of an EtherCalc "room"}

\item{ec_host}{See \code{\link[=ethercalc_host]{ethercalc_host()}}}
}
\description{
Open a browser window/tab to an EtherCalc "room"
}
