% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethercalc-package.R
\docType{package}
\name{ethercalc}
\alias{ethercalc}
\alias{ethercalc-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/ethercalc}
\item BugReports: \url{https://gitlab.com/hrbrmstr/ethercalc/issues}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/ethercalc}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/ethercalc/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
