% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethercalc-package.R
\docType{package}
\name{ethercalc}
\alias{ethercalc}
\alias{ethercalc-package}
\title{Orchestrate and Exchange Data with 'EtherCalc' Instances}
\description{
The 'EtherCalc' (\url{https://ethercalc.net/}) web application is a multi-user,
collaborative spreadsheet that can be run locally on virtually any modern operating system.
Methods are provided to orchestrate and exchange data with 'EtherCalc' instances.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/ethercalc}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/ethercalc/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
